[bugfix] set remote_src to yes and add integration test
fixes #3
