--- tests/integration/example/spec/default_spec.rb
@@ -1,22 +1,0 @@
-require "spec_helper"
-
-class ServiceNotReady < StandardError
-end
-
-sleep 10 if ENV["JENKINS_HOME"]
-
-context "after provisioning finished" do
-  describe server(:client1) do
-    it "should be able to ping server" do
-      result = current_server.ssh_exec("ping -c 1 #{server(:server1).server.address} && echo OK")
-      expect(result).to match(/OK/)
-    end
-  end
-
-  describe server(:server1) do
-    it "should be able to ping client" do
-      result = current_server.ssh_exec("ping -c 1 #{server(:client1).server.address} && echo OK")
-      expect(result).to match(/OK/)
-    end
-  end
-end
