+++ tests/integration/remote_src/.gitignore
