+++ tests/integration/remote_src/.rspec
