+++ tests/integration/remote_src/Vagrantfile
@@ -49,23 +49,6 @@ Vagrant.configure("2") do |config|
   end
   config.vm.box = "trombik/ansible-freebsd-10.3-amd64"
 
-  config.vm.define "client1" do |c|
-    c.vm.network "private_network", ip: "192.168.21.100"
-    c.vm.hostname = "client1.virtualbox.reallyenglish.com"
-    c.vm.provision :ansible do |ansible|
-      ansible.limit = "192.168.21.100"
-      ansible.playbook = "site.yml"
-      ansible.extra_vars = {
-        ansible_python_interpreter: "/usr/local/bin/python",
-        http_proxy: http_proxy,
-        https_proxy: http_proxy,
-        no_proxy: "localhost,127.0.0.1,.example.com"
-      }
-      ansible.inventory_path = "inventories/staging"
-      ansible.verbose = "v"
-    end
-  end
-
   config.vm.define "server1" do |c|
     c.vm.network "private_network", ip: "192.168.21.200"
     c.vm.hostname = "server1.virtualbox.reallyenglish.com"
