+++ tests/integration/remote_src/inventories/staging
@@ -1,9 +1,5 @@
-[client]
-192.168.21.100
-
 [server]
 192.168.21.200
 
 [all:children]
 server
-client
