+++ tests/integration/remote_src/server.yml
@@ -9,4 +9,4 @@
 
   pre_tasks:
   roles:
-    - packer-binary
+    - ansible-role-packer-binary
