+++ tests/integration/remote_src/site.yml
@@ -1,2 +1,1 @@
 - include: server.yml
-- include: client.yml
