+++ tests/integration/remote_src/spec/default_spec.rb
@@ -1,0 +1,15 @@
+require "spec_helper"
+
+class ServiceNotReady < StandardError
+end
+
+sleep 10 if ENV["JENKINS_HOME"]
+
+context "after provisioning finished" do
+  describe server(:server1) do
+    it "should be able to ping client" do
+      result = current_server.ssh_exec("echo -n OK")
+      expect(result).to eq "OK"
+    end
+  end
+end
